requirements.txt
fastapi==0.104.1
httpx==0.25.2
python-multipart==0.0.6
pyyaml==6.0.1
uvicorn==0.24.0
